+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
  
 ## Changelog
 
+### 2024-06-15
+- **Added**: Option to install just some specific GPT-RAG components: data ingestion, orchestrator and frontend.
+
 ### 2024-06-10
 - **Added**: Optional Reuse of Pre-Created Resources: You can now optionally reuse existing resources such as VNets, Azure OpenAI instances, etc.
-- **Added**: Ability to Reuse Application Service Environment.
 - **Added**: Use five subnets to align with the idealized architecture.
 - **Added**: Custom Addressing to allow specifying custom addressing for VNets and subnets, providing greater flexibility in network configuration.
 - **Added**: AI Integration Hub to leverage the power of various external data sources to enhance its capabilities. Currently, we have integrated the following products: